LPS-89140 Move comment packages to where they are used.
