+++ modules/apps/comment/comment-taglib/src/main/java/com/liferay/comment/taglib/internal/context/CommentDisplayContextProviderImpl.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.comment.display.context;
+package com.liferay.comment.taglib.internal.context;
 
-import com.liferay.portal.comment.display.context.util.DiscussionRequestHelper;
-import com.liferay.portal.comment.display.context.util.DiscussionTaglibHelper;
+import com.liferay.comment.taglib.internal.context.util.DiscussionRequestHelper;
+import com.liferay.comment.taglib.internal.context.util.DiscussionTaglibHelper;
 import com.liferay.portal.kernel.comment.Discussion;
 import com.liferay.portal.kernel.comment.DiscussionComment;
 import com.liferay.portal.kernel.comment.DiscussionPermission;
