+++ modules/apps/comment/comment-taglib/src/main/java/com/liferay/comment/taglib/internal/context/util/DiscussionTaglibHelper.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.comment.display.context.util;
+package com.liferay.comment.taglib.internal.context.util;
 
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.Validator;
