+++ modules/apps/comment/comment-taglib/src/main/resources/META-INF/resources/discussion/init.jsp
@@ -17,9 +17,9 @@
 <%@ include file="/init.jsp" %>
 
 <%@ page import="com.liferay.comment.constants.CommentConstants" %><%@
-page import="com.liferay.portal.comment.display.context.CommentDisplayContextProviderUtil" %><%@
-page import="com.liferay.portal.comment.display.context.util.DiscussionRequestHelper" %><%@
-page import="com.liferay.portal.comment.display.context.util.DiscussionTaglibHelper" %><%@
+page import="com.liferay.comment.taglib.internal.context.CommentDisplayContextProviderUtil" %><%@
+page import="com.liferay.comment.taglib.internal.context.util.DiscussionRequestHelper" %><%@
+page import="com.liferay.comment.taglib.internal.context.util.DiscussionTaglibHelper" %><%@
 page import="com.liferay.portal.kernel.comment.CommentManagerUtil" %><%@
 page import="com.liferay.portal.kernel.comment.Discussion" %><%@
 page import="com.liferay.portal.kernel.comment.DiscussionComment" %><%@
